+++ tests/Equinox.MemoryStore.Integration/MemoryStoreIntegration.fs
@@ -7,7 +7,7 @@ let createMemoryStore () =
     new VolatileStore()
 
 let createServiceMemory log store =
-    Backend.Cart.Service(log, MemoryResolver(store, Domain.Cart.Folds.fold, Domain.Cart.Folds.initial).Resolve)
+    Backend.Cart.Service(log, Resolver(store, Domain.Cart.Folds.fold, Domain.Cart.Folds.initial).Resolve)
 
 #nowarn "1182" // From hereon in, we may have some 'unused' privates (the tests)